Week 15/Week15 day2 - langgraph_tool_reuse/requirements.txt
langchain-core
langchain-text-splitters
langchain-community
langchain-openai
langgraph
faiss-cpu
pypdf
python-dotenv